+++ cedar-policy-core/src/est.rs
@@ -85,13 +85,13 @@ impl Policy {
     pub fn link(self, vals: &HashMap<ast::SlotId, EntityUidJson>) -> Result<Self, LinkingError> {
         Ok(Policy {
             effect: self.effect,
-            principal: self.principal.instantiate(vals)?,
-            action: self.action.instantiate(vals)?,
-            resource: self.resource.instantiate(vals)?,
+            principal: self.principal.link(vals)?,
+            action: self.action.link(vals)?,
+            resource: self.resource.link(vals)?,
             conditions: self
                 .conditions
                 .into_iter()
-                .map(|clause| clause.instantiate(vals))
+                .map(|clause| clause.link(vals))
                 .collect::<Result<Vec<_>, _>>()?,
             annotations: self.annotations,
         })
@@ -102,10 +102,7 @@ impl Clause {
     /// Fill in any slots in the clause using the values in `vals`. Throws an
     /// error if `vals` doesn't contain a necessary mapping, but does not throw
     /// an error if `vals` contains unused mappings.
-    pub fn instantiate(
-        self,
-        _vals: &HashMap<ast::SlotId, EntityUidJson>,
-    ) -> Result<Self, LinkingError> {
+    pub fn link(self, _vals: &HashMap<ast::SlotId, EntityUidJson>) -> Result<Self, LinkingError> {
         // currently, slots are not allowed in clauses
         Ok(self)
     }
@@ -2922,7 +2919,7 @@ mod test {
     }
 
     #[test]
-    fn instantiate() {
+    fn link() {
         let template = r#"
             permit(
                 principal == ?principal,
@@ -3893,7 +3890,7 @@ mod test {
         }
 
         #[test]
-        fn instantiate() {
+        fn link() {
             let template = r#"
             permit(
                 principal is User in ?principal,
@@ -3962,7 +3959,7 @@ mod test {
         }
 
         #[test]
-        fn instantiate_no_slot() {
+        fn link_no_slot() {
             let template = r#"permit(principal is User, action, resource is Doc);"#;
             let cst = parser::text_to_cst::parse_policy(template)
                 .unwrap()