+++ cedar-policy-validator/src/lib.rs
@@ -131,7 +131,6 @@ impl Validator {
             Some(
                 self.validate_entity_types(p)
                     .chain(self.validate_action_ids(p))
-                    .map(move |note| ValidationError::with_policy_id(p.id().clone(), None, note))
                     // We could usefully update this pass to apply to partial
                     // schema if it only failed when there is a known action
                     // applied to known principal/resource entity types that are
@@ -291,6 +290,8 @@ mod test {
                 Some("Action::\"action\"".to_string()),
             ),
         );
+
+        assert!(!result.validation_passed());
         assert!(result
             .validation_errors()
             .any(|x| x.error_kind() == principal_err.error_kind()));
@@ -305,7 +306,7 @@ mod test {
     }
 
     #[test]
-    fn top_level_validate_with_instantiations() -> Result<()> {
+    fn top_level_validate_with_links() -> Result<()> {
         let mut set = PolicySet::new();
         let schema: ValidatorSchema = serde_json::from_str::<SchemaFragment>(
             r#"
@@ -368,13 +369,14 @@ mod test {
             Vec::<&ValidationError>::new()
         );
 
-        // a valid instantiation is valid
+        // a valid link is valid
         let mut values = HashMap::new();
         values.insert(
             ast::SlotId::resource(),
             ast::EntityUID::from_components(
                 "some_namespace::Photo".parse().unwrap(),
                 ast::Eid::new("foo"),
+                None,
             ),
         );
         set.link(
@@ -386,13 +388,14 @@ mod test {
         let result = validator.validate(&set, ValidationMode::default());
         assert!(result.validation_passed());
 
-        // an invalid instantiation results in an error
+        // an invalid link results in an error
         let mut values = HashMap::new();
         values.insert(
             ast::SlotId::resource(),
             ast::EntityUID::from_components(
                 "some_namespace::Undefined".parse().unwrap(),
                 ast::Eid::new("foo"),
+                None,
             ),
         );
         set.link(
@@ -421,13 +424,14 @@ mod test {
         assert!(result.validation_errors().any(|x| x == &undefined_err));
         assert!(result.validation_errors().any(|x| x == &invalid_action_err));
 
-        // this is also an invalid instantiation (not a valid resource type for any action in the schema)
+        // this is also an invalid link (not a valid resource type for any action in the schema)
         let mut values = HashMap::new();
         values.insert(
             ast::SlotId::resource(),
             ast::EntityUID::from_components(
                 "some_namespace::User".parse().unwrap(),
                 ast::Eid::new("foo"),
+                None,
             ),
         );
         set.link(
@@ -446,7 +450,9 @@ mod test {
             loc.clone(),
             ValidationErrorKind::invalid_action_application(false, false),
         );
-        assert!(result.validation_errors().any(|x| x == &invalid_action_err));
+        assert!(result
+            .validation_errors()
+            .any(|x| x.error_kind() == invalid_action_err.error_kind()));
 
         Ok(())
     }