+++ cedar-policy-validator/src/typecheck/test_expr.rs
@@ -27,7 +27,7 @@ use smol_str::SmolStr;
 
 use crate::{
     type_error::TypeError, types::Type, AttributeAccess, AttributesOrContext, EntityType,
-    NamespaceDefinition, SchemaFragment, UnexpectedTypeHelp, ValidationMode,
+    LubContext, LubHelp, NamespaceDefinition, SchemaFragment, UnexpectedTypeHelp, ValidationMode,
 };
 
 use super::test_utils::{
@@ -90,7 +90,7 @@ fn slot_equals_typechecks() {
         shape: AttributesOrContext::default(),
     };
     // These don't typecheck in strict mode because the test_util expression
-    // typechecker doesn't have access to a schema, so it can't instantiate
+    // typechecker doesn't have access to a schema, so it can't link
     // the template slots with appropriate types. Similar policies that pass
     // strict typechecking are in the test_policy file.
     let schema = NamespaceDefinition::new([("typename".parse().unwrap(), etype)], []);
@@ -144,6 +144,8 @@ fn heterogeneous_set() {
         vec![TypeError::incompatible_types(
             set,
             vec![Type::singleton_boolean(true), Type::primitive_long()],
+            LubHelp::None,
+            LubContext::Set,
         )],
     );
 }
@@ -671,6 +673,8 @@ fn record_get_attr_lub_typecheck_fails() {
                 )]),
                 Type::primitive_long(),
             ],
+            LubHelp::None,
+            LubContext::Conditional,
         )],
     );
 }
@@ -1025,6 +1029,8 @@ fn if_no_lub_error() {
         vec![TypeError::incompatible_types(
             if_expr,
             vec![Type::primitive_long(), Type::primitive_string()],
+            LubHelp::None,
+            LubContext::Conditional,
         )],
     );
 }
@@ -1038,6 +1044,8 @@ fn if_typecheck_fails() {
             TypeError::incompatible_types(
                 if_expr,
                 vec![Type::primitive_long(), Type::primitive_string()],
+                LubHelp::None,
+                LubContext::Conditional,
             ),
             TypeError::expected_type(
                 Expr::val("fail"),