+++ cedar-policy-validator/src/typecheck/test_strict.rs
@@ -35,7 +35,8 @@ use cedar_policy_core::{
 use crate::{
     typecheck::test_utils::assert_policy_typecheck_fails,
     types::{AttributeType, EffectSet, OpenTag, RequestEnv, Type},
-    IncompatibleTypes, SchemaFragment, TypeError, TypeErrorKind, ValidationMode,
+    IncompatibleTypes, LubContext, LubHelp, SchemaFragment, TypeError, TypeErrorKind,
+    ValidationMode,
 };
 
 use super::test_utils::with_typechecker_from_schema;
@@ -99,6 +100,8 @@ fn assert_types_must_match(
     e_strict: Expr,
     expected_type: Type,
     unequal_types: impl IntoIterator<Item = Type>,
+    hint: LubHelp,
+    context: LubContext,
 ) {
     assert_strict_type_error(
         schema,
@@ -108,6 +111,8 @@ fn assert_types_must_match(
         expected_type,
         TypeErrorKind::IncompatibleTypes(IncompatibleTypes {
             types: unequal_types.into_iter().collect(),
+            hint,
+            context,
         }),
     )
 }
@@ -228,6 +233,8 @@ fn eq_strict_types_mismatch() {
             Expr::from_str(r#"1 == "foo""#).unwrap(),
             Type::primitive_boolean(),
             [Type::primitive_string(), Type::primitive_long()],
+            LubHelp::None,
+            LubContext::Equality,
         )
     })
 }
@@ -242,6 +249,8 @@ fn contains_strict_types_mismatch() {
             Expr::from_str(r#"[1].contains("test")"#).unwrap(),
             Type::primitive_boolean(),
             [Type::primitive_long(), Type::primitive_string()],
+            LubHelp::None,
+            LubContext::Contains,
         )
     })
 }
@@ -259,6 +268,8 @@ fn contains_any_strict_types_mismatch() {
                 Type::set(Type::named_entity_reference_from_str("User")),
                 Type::set(Type::primitive_long()),
             ],
+            LubHelp::None,
+            LubContext::ContainsAnyAll,
         )
     })
 }
@@ -276,6 +287,8 @@ fn contains_all_strict_types_mismatch() {
                 Type::set(Type::named_entity_reference_from_str("User")),
                 Type::set(Type::primitive_long()),
             ],
+            LubHelp::None,
+            LubContext::ContainsAnyAll,
         )
     })
 }
@@ -338,6 +351,8 @@ fn if_bool_strict_type_mismatch() {
                 Type::named_entity_reference_from_str("User"),
                 Type::named_entity_reference_from_str("Photo"),
             ],
+            LubHelp::EntityType,
+            LubContext::Conditional,
         )
     })
 }
@@ -355,6 +370,8 @@ fn set_strict_types_mismatch() {
                 Type::named_entity_reference_from_str("User"),
                 Type::named_entity_reference_from_str("Photo"),
             ],
+            LubHelp::EntityType,
+            LubContext::Set,
         )
     })
 }
@@ -422,6 +439,8 @@ fn entity_in_lub() {
                 Type::named_entity_reference_from_str("User"),
                 Type::named_entity_reference_from_str("Photo"),
             ],
+            LubHelp::EntityType,
+            LubContext::Conditional,
         )
     });
 }
@@ -447,6 +466,8 @@ fn test_and() {
             Expr::from_str(r#"(1 == (2 > 0)) && true"#).unwrap(),
             Type::primitive_boolean(),
             [Type::primitive_long(), Type::primitive_boolean()],
+            LubHelp::None,
+            LubContext::Equality,
         );
         assert_types_must_match(
             s,
@@ -455,6 +476,8 @@ fn test_and() {
             Expr::from_str(r#"true && (1 == (2 > 0))"#).unwrap(),
             Type::primitive_boolean(),
             [Type::primitive_long(), Type::primitive_boolean()],
+            LubHelp::None,
+            LubContext::Equality,
         );
     })
 }
@@ -476,6 +499,8 @@ fn test_or() {
             Expr::from_str(r#"(1 == (2 > 0)) || false"#).unwrap(),
             Type::primitive_boolean(),
             [Type::primitive_boolean(), Type::primitive_long()],
+            LubHelp::None,
+            LubContext::Equality,
         );
         assert_types_must_match(
             s,
@@ -484,6 +509,8 @@ fn test_or() {
             Expr::from_str(r#"false || (1 == (2 > 0))"#).unwrap(),
             Type::primitive_boolean(),
             [Type::primitive_boolean(), Type::primitive_long()],
+            LubHelp::None,
+            LubContext::Equality,
         );
     })
 }
@@ -505,6 +532,8 @@ fn test_unary() {
             Expr::from_str(r#"!(1 == "foo")"#).unwrap(),
             Type::primitive_boolean(),
             [Type::primitive_long(), Type::primitive_string()],
+            LubHelp::None,
+            LubContext::Equality,
         );
     })
 }
@@ -526,6 +555,8 @@ fn test_mul() {
             Expr::from_str(r#"2*(if 1 == false then 3 else 4)"#).unwrap(),
             Type::primitive_long(),
             [Type::primitive_long(), Type::singleton_boolean(false)],
+            LubHelp::None,
+            LubContext::Equality,
         );
     })
 }
@@ -547,6 +578,8 @@ fn test_like() {
             Expr::from_str(r#"(if 1 == false then "foo" else "bar") like "bar""#).unwrap(),
             Type::primitive_boolean(),
             [Type::primitive_long(), Type::singleton_boolean(false)],
+            LubHelp::None,
+            LubContext::Equality,
         );
     })
 }
@@ -568,6 +601,8 @@ fn test_get_attr() {
             Expr::from_str(r#"{name: 1 == "foo"}.name"#).unwrap(),
             Type::primitive_boolean(),
             [Type::primitive_long(), Type::primitive_string()],
+            LubHelp::None,
+            LubContext::Equality,
         );
     })
 }
@@ -605,6 +640,8 @@ fn test_has_attr() {
                     Type::primitive_long(),
                 )]),
             ],
+            LubHelp::RecordWidth,
+            LubContext::Conditional,
         );
     })
 }
@@ -626,7 +663,9 @@ fn test_extension() {
             Expr::from_str(r#"ip("192.168.1.0/8").isInRange(if 1 == false then ip("127.0.0.1") else ip("192.168.1.1"))"#).unwrap(),
             Expr::from_str(r#"ip("192.168.1.0/8").isInRange(if 1 == false then ip("127.0.0.1") else ip("192.168.1.1"))"#).unwrap(),
             Type::primitive_boolean(),
-            [Type::primitive_long(), Type::singleton_boolean(false)]
+            [Type::primitive_long(), Type::singleton_boolean(false)],
+            LubHelp::None,
+            LubContext::Equality,
         );
     })
 }
@@ -716,6 +755,8 @@ fn qualified_record_attr() {
                     OpenTag::ClosedAttributes,
                 ),
             ],
+            LubHelp::AttributeQualifier,
+            LubContext::Equality,
         )],
     );
 }